moved sourceControlList to metadata endpoint
